# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meteora DLMM swap indexer API",
    "version": "0.1.0",
    "description": "Read-only REST API over indexed Meteora DLMM swaps, except `POST /v1/backfills` and `DELETE /v1/backfills/{job_id}`, which request and cancel history fills. Times are UTC RFC 3339, and token amounts are strings in raw base units and in human units so that no precision is lost."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "paths": {
...
    "/v1/pools": {
      "get": {
        "operationId": "listPools",
        "summary": "Pools ranked by 24-hour USD volume.",
        "description": "Offset pages over every indexed pool, ordered by `volume_usd_24h` descending (a pool without priced volume counts as 0), then `swap_count_24h` descending, then address. The ranking moves as blocks land, so a pool can cross a page boundary between two requests. An offset past the end gives an empty page with the same `page.total`.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size, 1 to 100. Canonical decimal only: no sign, no leading zero, no spaces.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Number of ranked pools to skip. Canonical decimal only.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 4294967295,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of pools.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoolList"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Codes: `invalid_query`: The query string cannot be parsed. `invalid_limit`: `limit` is not a canonical decimal integer from 1 to 100. `invalid_offset`: `offset` is not a canonical non-negative decimal integer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error. Codes: `internal`: An unexpected server fault. The server log has the cause.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable. Codes: `projection_rebuilding`: The pool ranking reads `pool_stats` and `pool_volume_1h`; either one mid-rebuild gives 503. `database_unavailable`: The database did not answer. Retry later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pools/{pool}": {
      "get": {
        "operationId": "getPool",
        "summary": "One pool's summary.",
        "description": "The same object as an entry of `GET /v1/pools`.",
        "parameters": [
          {
            "name": "pool",
            "in": "path",
            "required": true,
            "description": "Pool (DLMM `lb_pair`) address, base58.",
            "schema": {
              "type": "string",
              "pattern": "^[1-9A-HJ-NP-Za-km-z]{32,44}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The pool summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pool"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Codes: `invalid_pool`: The `pool` path segment is not a base58 32-byte address.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found. Codes: `pool_not_found`: The pool is well formed but the indexer has never seen it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error. Codes: `internal`: An unexpected server fault. The server log has the cause.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable. Codes: `projection_rebuilding`: The summary reads `pool_stats` and `pool_volume_1h`; either one mid-rebuild gives 503. An unknown pool is still 404. `database_unavailable`: The database did not answer. Retry later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pools/{pool}/volume": {
      "get": {
        "operationId": "getPoolVolume",
        "summary": "Hourly or daily volume of one pool.",
        "description": "The API floors `from` and ceils `to` to UTC bucket boundaries (top of the hour, or UTC midnight) and echoes the aligned window. Every bucket in the window is present, oldest first; a bucket without swaps has zero volumes and a null `volume_usd`. At most 744 hourly or 366 daily buckets per request. Validation order: pool, existence, bucket, from, to, alignment, cap.",
        "parameters": [
          {
            "name": "pool",
            "in": "path",
            "required": true,
            "description": "Pool (DLMM `lb_pair`) address, base58.",
            "schema": {
              "type": "string",
              "pattern": "^[1-9A-HJ-NP-Za-km-z]{32,44}$"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "required": true,
            "description": "Bucket size.",
            "schema": {
              "type": "string",
              "enum": [
                "hour",
                "day"
              ]
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "description": "Start instant, RFC 3339 with any offset (encode `+` as `%2B`). Floored to the bucket.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "description": "End instant, RFC 3339 with any offset. Ceiled to the bucket; the aligned end is exclusive.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The pool's buckets over the aligned window.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Volume"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Codes: `invalid_pool`: The `pool` path segment is not a base58 32-byte address. `invalid_query`: The query string cannot be parsed. `invalid_bucket`: `bucket` is missing or is not `hour` or `day`. `invalid_time`: `from` or `to` is missing or is not an RFC 3339 timestamp. The message names the parameter. `invalid_range`: `from` is not before `to` after alignment to the bucket. `range_too_large`: The aligned range holds more than 744 hourly or 366 daily buckets. The message gives the count and the cap.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found. Codes: `pool_not_found`: The pool is well formed but the indexer has never seen it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error. Codes: `internal`: An unexpected server fault. The server log has the cause.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable. Codes: `projection_rebuilding`: `bucket=hour` reads `pool_volume_1h` and `bucket=day` reads `pool_volume_1d`; only the table of the requested bucket gives 503. `database_unavailable`: The database did not answer. Retry later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pools/{pool}/swaps": {
      "get": {
        "operationId": "listPoolSwaps",
        "summary": "One pool's swaps, newest first.",
        "description": "Keyset pages, newest first by block time and log position. Swaps that land while a client pages never shift or repeat a page. Served from the swap log, so a projection rebuild never gives 503 here.",
        "parameters": [
          {
            "name": "pool",
            "in": "path",
            "required": true,
            "description": "Pool (DLMM `lb_pair`) address, base58.",
            "schema": {
              "type": "string",
              "pattern": "^[1-9A-HJ-NP-Za-km-z]{32,44}$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size, 1 to 100. Canonical decimal only: no sign, no leading zero, no spaces.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "description": "Opaque cursor: the `page.next_cursor` of the previous page. Omit it for the newest page. Clients must not build or edit it.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of swaps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwapList"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Codes: `invalid_pool`: The `pool` path segment is not a base58 32-byte address. `invalid_query`: The query string cannot be parsed. `invalid_limit`: `limit` is not a canonical decimal integer from 1 to 100. `invalid_cursor`: `before` is not a `next_cursor` that this API issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found. Codes: `pool_not_found`: The pool is well formed but the indexer has never seen it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal error. Codes: `internal`: An unexpected server fault. The server log has the cause.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable. Codes: `database_unavailable`: The database did not answer. Retry later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "description": "Every non-2xx response with a body.",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "Pool": {
        "type": "object",
        "description": "One pool's summary. The 24-hour window is the current UTC hour and the 23 hours before it.",
        "additionalProperties": false,
        "required": [
          "address",
          "mint_x",
          "mint_y",
          "decimals_x",
          "decimals_y",
          "swap_count_24h",
          "volume_usd_24h",
          "first_swap_at",
          "last_swap_at"
        ],
        "properties": {
          "address": {
            "type": "string",
            "pattern": "^[1-9A-HJ-NP-Za-km-z]{32,44}$",
            "description": "Base58 32-byte address. Pool (DLMM `lb_pair`) address."
          },
          "mint_x": {
            "type": "string",
            "pattern": "^[1-9A-HJ-NP-Za-km-z]{32,44}$",
            "description": "Base58 32-byte address. Mint of the pool's token X."
          },
          "mint_y": {
            "type": "string",
            "pattern": "^[1-9A-HJ-NP-Za-km-z]{32,44}$",
            "description": "Base58 32-byte address. Mint of the pool's token Y."
          },
          "decimals_x": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Decimals of token X. Null until the indexer fetched the mint.",
            "minimum": 0
          },
          "decimals_y": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Decimals of token Y. Null until the indexer fetched the mint.",
            "minimum": 0
          },
          "swap_count_24h": {
            "type": "integer",
            "description": "Swaps in the 24-hour window.",
            "minimum": 0
          },
          "volume_usd_24h": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^[0-9]+(\\.[0-9]+)?$",
            "description": "Exact decimal string, no exponent, trailing zeros removed. USD volume of the priced swaps in the 24-hour window. Null when the window holds no priced swap (no swaps, or every swap unpriced)."
          },
          "first_swap_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}Z$",
            "description": "RFC 3339 instant in UTC, whole seconds, `Z` suffix. Block time of the pool's first indexed swap. Null before the first swap."
          },
          "last_swap_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}Z$",
            "description": "RFC 3339 instant in UTC, whole seconds, `Z` suffix. Block time of the pool's last indexed swap. Null before the first swap."
          }
        }
      },
...
      "PoolList": {
        "type": "object",
        "description": "One page of ranked pools.",
        "additionalProperties": false,
        "required": [
          "pools",
          "page"
        ],
        "properties": {
          "pools": {
            "type": "array",
            "description": "Pools in rank order. Empty past the end.",
            "items": {
              "$ref": "#/components/schemas/Pool"
            }
          },
          "page": {
            "$ref": "#/components/schemas/OffsetPage"
          }
        }
      },
...
      "Volume": {
        "type": "object",
        "description": "A pool's buckets over the aligned window.",
        "additionalProperties": false,
        "required": [
          "pool",
          "mint_x",
          "mint_y",
          "decimals_x",
          "decimals_y",
          "first_swap_at",
          "bucket",
          "from",
          "to",
          "buckets"
        ],
        "properties": {
          "pool": {
            "type": "string",
            "pattern": "^[1-9A-HJ-NP-Za-km-z]{32,44}$",
            "description": "Base58 32-byte address. Pool address."
          },
          "mint_x": {
            "type": "string",
            "pattern": "^[1-9A-HJ-NP-Za-km-z]{32,44}$",
            "description": "Base58 32-byte address. Mint of token X."
          },
          "mint_y": {
            "type": "string",
            "pattern": "^[1-9A-HJ-NP-Za-km-z]{32,44}$",
            "description": "Base58 32-byte address. Mint of token Y."
          },
          "decimals_x": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Decimals of token X. Null until fetched.",
            "minimum": 0
          },
          "decimals_y": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Decimals of token Y. Null until fetched.",
            "minimum": 0
          },
          "first_swap_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}Z$",
            "description": "RFC 3339 instant in UTC, whole seconds, `Z` suffix. Block time of the pool's first indexed swap. Buckets before it read zero because nothing was indexed, not because nothing traded. Null before the first swap and while `pool_stats` rebuilds."
          },
          "bucket": {
            "type": "string",
            "enum": [
              "hour",
              "day"
            ],
            "description": "Bucket size that applied."
          },
          "from": {
            "type": "string",
            "format": "date-time",
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}Z$",
            "description": "RFC 3339 instant in UTC, whole seconds, `Z` suffix. Aligned window start, inclusive."
          },
          "to": {
            "type": "string",
            "format": "date-time",
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}Z$",
            "description": "RFC 3339 instant in UTC, whole seconds, `Z` suffix. Aligned window end, exclusive."
          },
          "buckets": {
            "type": "array",
            "minItems": 1,
            "description": "Every bucket from `from` to `to`, oldest first, with none missing.",
            "items": {
              "$ref": "#/components/schemas/VolumeBucket"
            }
          }
        }
      },
...
    }
  }
}